# client/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Financial Document Analyzer API",
      "description": "FastAPI application with JWT auth, document management, and CrewAI agents",
      "version": "2.0.0"
    },
    "paths": {
...
      "/documents/": {
        "get": {
          "tags": [
            "documents"
          ],
          "summary": "List Documents",
          "description": "List documents for the current user with optional search",
          "operationId": "list_documents_documents__get",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "q",
              "in": "query",
              "required": false,
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Q"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/DocumentMetadata"
                    },
                    "title": "Response List Documents Documents  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/documents/upload": {
        "post": {
          "tags": [
            "documents"
          ],
          "summary": "Upload Document",
          "description": "Upload a new document",
          "operationId": "upload_document_documents_upload_post",
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/Body_upload_document_documents_upload_post"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/DocumentMetadata"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          },
          "security": [
            {
              "HTTPBearer": []
            }
          ]
        }
      },
      "/documents/download/{filename}": {
        "get": {
          "tags": [
            "documents"
          ],
          "summary": "Download Document",
          "description": "Download a document by filename",
          "operationId": "download_document_documents_download__filename__get",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "filename",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Filename"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/documents/{document_id}": {
        "delete": {
          "tags": [
            "documents"
          ],
          "summary": "Delete Document",
          "description": "Delete a document by document ID",
          "operationId": "delete_document_documents__document_id__delete",
          "security": [
            {
              "HTTPBearer": []
            }
          ],
          "parameters": [
            {
              "name": "document_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Document Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Body_upload_document_documents_upload_post": {
          "properties": {
            "file": {
              "type": "string",
              "format": "binary",
              "title": "File"
            },
            "name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          "type": "object",
          "required": [
            "file"
          ],
          "title": "Body_upload_document_documents_upload_post"
        },
...
        "DocumentMetadata": {
          "properties": {
            "id": {
              "type": "string",
              "title": "Id"
            },
            "name": {
              "type": "string",
              "title": "Name"
            },
            "size_bytes": {
              "type": "integer",
              "title": "Size Bytes"
            },
            "modified_at": {
              "type": "string",
              "title": "Modified At"
            },
            "path": {
              "type": "string",
              "title": "Path"
            },
            "download_url": {
              "type": "string",
              "title": "Download Url"
            }
          },
          "type": "object",
          "required": [
            "id",
            "name",
            "size_bytes",
            "modified_at",
            "path",
            "download_url"
          ],
          "title": "DocumentMetadata"
        },
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      },
      "securitySchemes": {
        "HTTPBearer": {
          "type": "http",
          "scheme": "bearer"
        }
      }
    }
  }